requirements.txt
aniso8601==10.0.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
Werkzeug==3.1.3
Flask-Migrate==4.0.5
python-dotenv==1.0.1
